Shifted actor portrait API and runtime content to its own space.
