+++ src/api/actor-portrait/ActorPortraitApi.ts
@@ -1,0 +1,8 @@
+import type { PortraitContextMenuCommand } from '..';
+import { ActorPortraitRuntime } from 'src/runtime/ActorPortraitRuntime';
+
+export class ActorPortraitApi {
+  registerContextMenuCommands(commands: PortraitContextMenuCommand[]) {
+    ActorPortraitRuntime.registerPortraitContextMenuCommands([...commands]);
+  }
+}
